HSTA_MER-main/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
asyncio==3.4.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decord==0.6.0
et-xmlfile==1.1.0
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.2
hjson==3.1.0
huggingface-hub==0.18.0
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
ipython-genutils==0.2.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
libaio==0.9.2
Markdown==3.5.1
matplotlib==3.7.3
mpmath==1.3.0
nbconvert==5.6.1
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
pnvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.1
prefetch-generator==1.0.3
protobuf==4.25.0
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pynvml==11.5.0
pyparsing==3.1.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==19.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shutilwhich==1.1.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.4.12
torch==1.13.1+cu117
torchdiffeq==0.2.3
torchsampler==0.1.2
torchvision==0.14.1+cu117
tqdm==4.66.1
triton==2.1.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
